Prevent the 'no handlers' message from logging
When using recurrent with Python 2.7, the logging module will print
the message 'No handlers could be found for logger "recurrent"'.
This occurs when not handlers have been set up in the logging module,
which is the case when importing recurrent in an interpreter, for
example.

This commit adds a NullHandler, which does nothing when other handlers
exist, and suppresses the message if no other handlers exist.